Backend/requirements.txt
Flask
flask-cors
google-generativeai
googletrans==4.0.0-rc1
